Time: 4 ms (31.17%) | Memory: 6.5 MB (46.32%) - LeetSync

+++ 2824-check-if-the-number-is-fascinating/check-if-the-number-is-fascinating.cpp
@@ -3,7 +3,7 @@ public:
     bool isFascinating(int n) {
        string a=to_string(n)+to_string(2*n)+to_string(3*n);
        
-       long long ans=stoll(a);
+       long ans=stoll(a);
        int multi=1;
         while(ans!=0){
           multi=multi*(ans%10);
